refactor(chat): extract ChatInput.Attach to its own file (LOC ceiling)
Splitting the composer's + attachment control into chat-input-attach.tsx keeps
chat-composer.tsx under its ban-chat-antipatterns LOC ceiling (338 < 389) after
the single-action + attachments-inside-card changes. Public API unchanged:
ChatInputAttach is re-exported and still wired as ChatInput.Attach.
